+++ tests/test_board.py
@@ -1,17 +1,87 @@
 import unittest
 from game.board import Board
+from game.models import Tile
 
 class TestBoard(unittest.TestCase):
     def test_init(self):
         board = Board()
+        self.assertEqual(len(board.grid), 15)
         self.assertEqual(
-            len(board.grid),
-            15,
-        )
-        self.assertEqual(
-            len(board.grid[0]),
-            15,
-        )
+            len(board.grid[0]),15)
+        
+    def test_word_inside_board(self):
+        board = Board()
+        word = "Facultad"
+        location = (5, 4)
+        orientation = "H"
+        word_is_valid = board.validate_word_inside_board(word, location, orientation)
+        assert word_is_valid == True
+    
+    def test_word_out_of_board(self):
+        board = Board()
+        word = "Facultad"
+        location = (14, 4)
+        orientation = "H"
+        word_is_valid = board.validate_word_inside_board(word, location, orientation)
+        assert word_is_valid == False
 
+    def test_board_is_empty(self):
+        board = Board()
+        self.assertEqual(board.isEmpty(), True)
+
+    def test_board_is_not_empty(self):
+        board = Board()
+        board.addTile(7, 7, Tile('C', 1))
+        board.addTile(8, 7, Tile('A', 1))
+        board.addTile(9, 7, Tile('S', 1))
+        board.addTile(10, 7, Tile('A', 1))
+        self.assertEqual(board.isEmpty(), False)
+
+    def test_place_word_empty_board_horizontal_fine(self):
+        board = Board()
+        word = "Facultad"
+        location = (7, 4)
+        orientation = "H"
+        word_is_valid = board.validate_word_inside_board(word, location, orientation)
+        assert word_is_valid == True
+
+    """
+    def test_place_word_empty_board_horizontal_wrong(self):
+        board = Board()
+        word = "Facultad"
+        location = (2, 4)
+        orientation = "H"
+        word_is_valid = board.validate_word_inside_board(word, location, orientation)
+        assert word_is_valid == False
+    """
+    def test_place_word_empty_board_vertical_fine(self):
+        board = Board()
+        word = "Facultad"
+        location = (4, 7)
+        orientation = "V"
+        word_is_valid = board.validate_word_inside_board(word, location, orientation)
+        assert word_is_valid == True
+    
+    """
+    def test_place_word_empty_board_vertical_wrong(self):
+        board = Board()
+        word = "Facultad"
+        location = (2, 4)
+        orientation = "V"
+        word_is_valid = board.validate_word_inside_board(word, location, orientation)
+        assert word_is_valid == False
+
+    def test_place_word_not_empty_board_horizontal_fine(self):
+        board = Board()
+        board.addTile(7, 7, Tile('C', 1))
+        board.addTile(8, 7, Tile('A', 1))
+        board.addTile(9, 7, Tile('S', 1))
+        board.addTile(10, 7, Tile('A', 1))
+        word = "Facultad"
+        location = (8, 6)
+        orientation = "H"
+        word_is_valid = board.validate_word_inside_board(word, location, orientation)
+        assert word_is_valid == True
+    """
 if __name__ == '__main__':
     unittest.main()