+++ tests/test_cell.py
@@ -7,26 +7,26 @@ class TestCell(unittest.TestCase):
         cell = Cell(multiplier=2, multiplier_type='letter')
         self.assertEqual(cell.multiplier,2,)
         self.assertEqual(cell.multiplier_type,'letter')
-        self.assertIsNone(cell.letter)
+        self.assertIsNone(cell.tile)
         self.assertEqual(cell.calculate_value(), 0)
 
     def test_add_letter(self):
         cell = Cell(multiplier=1, multiplier_type='')
-        letter = Tile(letter='p', value=3)
-        cell.add_letter(letter=letter)
-        self.assertEqual(cell.letter, letter)
+        add = Tile('P', 3)
+        cell.addValue(add)
+        self.assertEqual(cell.tile, add)
 
     def test_cell_value(self):
         cell = Cell(multiplier=2, multiplier_type='letter')
-        letter = Tile(letter='p', value=3)
-        cell.add_letter(letter=letter)
+        add = Tile('P', 3)
+        cell.addValue(add)
         self.assertEqual(cell.calculate_value(), 6)
 
     def test_cell_multiplier_word(self):
         cell = Cell(multiplier=2, multiplier_type='word')
-        letter = Tile(letter='p', value=3)
-        cell.add_letter(letter=letter)
+        cell.addValue(Tile('P', 3))
         self.assertEqual(cell.calculate_value(), 3)
+
 
 if __name__ == '__main__':
     unittest.main()