+++ CHANGELOG.md
@@ -6,10 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added [04/10/23]
+### Added [05/10/23]
+- functions (addTile, isEmpty, addValue), tests for test_board.py
+### Added [04/10/23] 
 - more tests for scrabble_game
 - word validation, parsing and putting [logic missing, init idea]
 - cli.py, test_cli.py
+- tests for test_board
+- board functionality [validate word inside board, is empty]
 ### Added [03/10/23]
 - tests for letters.py
 - tests for test_player.py, logic for letter checking in player.py